Retrieve doctors by postal code
When querying for just no parameters, despite spitting out ~2622 results on 105
Pages, there were some missing doctors. To fix that, one could just
fetch each and every `postleitzal`. This is not as performant as
fetching them all (There are at 224 more requests to the search) but at
least we get a comprehensive result.